Change service entity id type back to int

+++ internal/services/auth_service.go
@@ -17,7 +17,7 @@ func NewAuthService(userRepo repositories.UserRepository) *AuthService {
 	return &AuthService{userRepo: userRepo}
 }
 
-func (s *AuthService) Login(f *dtos.UserLoginForm) (uint64, error) {
+func (s *AuthService) Login(f *dtos.UserLoginForm) (int, error) {
 	f.Validator.Check(v.NotBlank(f.Username), "username", "Username should not be blank")
 
 	userFromDb, err := s.userRepo.GetByUsername(f.Username)
